requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
faiss-cpu==1.7.4
filelock==3.20.0
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.16.4
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lit==18.1.8
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.0
packaging==25.0
pillow==12.0.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
transformers==4.32.1
triton==2.0.0
typing_extensions==4.15.0
urllib3==2.5.0
